gmoccapy: fix wrong states of spindle button images after program abort
When a program with spindle on is aborted (assuming spindle off before program start), the images of the spindle buttons get a wrong state: the CSS property "-gtk-icon-effect" is set to "dim" which should be only set when they are disabled.
Toggling the sensitive property twice corrects this wrong CSS property.